+++ publish/changeLog.md
@@ -1,3 +1,3 @@
-### Fix
+### Add
 
-- remove log output
+- add music comment